\name{bnpmr-package}
\alias{bnpmr-package}
\alias{getPrior}
\alias{getProp}
\alias{checkArgs}
\docType{package}
\title{
Monotonic Regression
}
\description{
  Implements the Bayesian monotonic regression method described in
  Bornkamp and Ickstadt (2009)
}
\details{
\tabular{ll}{
Package: \tab bnpmr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-25\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Bjoern Bornkamp

Maintainer: Bjoern Bornkamp <bbnkmp@gmail.com>

}
\references{
  Bornkamp, B., Ickstadt, K. (2009). Bayesian nonparametric
  estimation of continuous monotone functions with applications to
  dose-response analysis. Biometrics, 65, 198-205.
}
\keyword{ package }
\seealso{
  \code{\link{pred.bnpmr}}, \code{\link{bnpmr}}
}
